feat: merge config type of plugin-proxy and plugin-unbundle
