+++ packages/cli/core/src/config/index.ts
@@ -1,4 +1,3 @@
-import http from 'http';
 import { loadConfig } from '@modern-js/load-config';
 import Ajv, { ErrorObject } from 'ajv';
 import ajvKeywords from 'ajv-keywords';
@@ -7,193 +6,24 @@ import {
   createDebugger,
   getPort,
   isDev,
-  MetaOptions,
   PLUGIN_SCHEMAS,
   chalk,
 } from '@modern-js/utils';
 import mergeWith from 'lodash.mergewith';
 import betterAjvErrors from 'better-ajv-errors';
 import { codeFrameColumns } from '@babel/code-frame';
-import type { NextFunction, ProxyOptions } from '@modern-js/types';
-import { PluginConfig } from '../loadPlugins';
+
 import { repeatKeyWarning } from '../utils/repeatKeyWarning';
 import { defaults } from './defaults';
 import { mergeConfig, NormalizedConfig } from './mergeConfig';
 import { patchSchema, PluginValidateSchema } from './schema';
+import type { UserConfig, ConfigParam, LoadedConfig } from './types';
 
 const debug = createDebugger('resolve-config');
 
 export { defaults as defaultsConfig };
 export { mergeConfig };
-
-interface SourceConfig {
-  entries?: Record<
-    string,
-    | string
-    | {
-        entry: string;
-        enableFileSystemRoutes?: boolean;
-        disableMount?: boolean;
-      }
-  >;
-  disableDefaultEntries?: boolean;
-  entriesDir?: string;
-  configDir?: string;
-  apiDir?: string;
-  envVars?: Array<string>;
-  globalVars?: Record<string, string>;
-  alias?:
-    | Record<string, string>
-    | ((aliases: Record<string, string>) => Record<string, unknown>);
-  moduleScopes?:
-    | Array<string | RegExp>
-    | ((scopes: Array<string | RegExp>) => void)
-    | ((scopes: Array<string | RegExp>) => Array<string | RegExp>);
-  include?: Array<string | RegExp>;
-}
-
-interface OutputConfig {
-  assetPrefix?: string;
-  htmlPath?: string;
-  jsPath?: string;
-  cssPath?: string;
-  mediaPath?: string;
-  path?: string;
-  title?: string;
-  titleByEntries?: Record<string, string>;
-  meta?: MetaOptions;
-  metaByEntries?: Record<string, MetaOptions>;
-  inject?: 'body' | 'head' | boolean;
-  injectByEntries?: Record<string, 'body' | 'head' | boolean>;
-  mountId?: string;
-  favicon?: string;
-  faviconByEntries?: Record<string, string | undefined>;
-  copy?: Array<Record<string, unknown> & { from: string }>;
-  scriptExt?: Record<string, unknown>;
-  disableTsChecker?: boolean;
-  disableHtmlFolder?: boolean;
-  disableCssModuleExtension?: boolean;
-  disableCssExtract?: boolean;
-  enableCssModuleTSDeclaration?: boolean;
-  disableMinimize?: boolean;
-  enableInlineStyles?: boolean;
-  enableInlineScripts?: boolean;
-  disableSourceMap?: boolean;
-  disableInlineRuntimeChunk?: boolean;
-  disableAssetsCache?: boolean;
-  enableLatestDecorators?: boolean;
-  polyfill?: 'off' | 'usage' | 'entry' | 'ua';
-  dataUriLimit?: number;
-  templateParameters?: Record<string, unknown>;
-  templateParametersByEntries?: Record<
-    string,
-    Record<string, unknown> | undefined
-  >;
-  cssModuleLocalIdentName?: string;
-  enableModernMode?: boolean;
-  federation?: boolean;
-  disableNodePolyfill?: boolean;
-  enableTsLoader?: boolean;
-}
-
-interface ServerConfig {
-  routes?: Record<
-    string,
-    | string
-    | string[]
-    | {
-        route: string | string[];
-        disableSpa?: boolean;
-      }
-  >;
-  publicRoutes?: { [filepath: string]: string };
-  ssr?: boolean | Record<string, unknown>;
-  ssrByEntries?: Record<string, boolean | Record<string, unknown>>;
-  baseUrl?: string | Array<string>;
-  port?: number;
-  logger?: boolean | Record<string, any>;
-  metrics?: boolean | Record<string, any>;
-  enableMicroFrontendDebug?: boolean;
-}
-
-interface DevConfig {
-  assetPrefix?: string | boolean;
-  https?: boolean;
-}
-
-interface MicroFrontend {
-  enableHtmlEntry?: boolean;
-  externalBasicLibrary?: boolean;
-  moduleApp?: string;
-}
-
-interface DeployConfig {
-  microFrontend?: false | MicroFrontend;
-  domain?: string | Array<string>;
-  domainByEntries?: Record<string, string | Array<string>>;
-}
-
-type ConfigFunction =
-  | Record<string, unknown>
-  | ((config: Record<string, unknown>) => Record<string, unknown> | void);
-
-type RequestHandler = (
-  req: http.IncomingMessage,
-  res: http.ServerResponse,
-  next: NextFunction,
-) => void;
-type DevServerConfig = {
-  proxy?: ProxyOptions;
-  headers?: Record<string, string>;
-  before?: RequestHandler[];
-  after?: RequestHandler[];
-  [propsName: string]: any;
-};
-
-interface ToolsConfig {
-  webpack?: ConfigFunction;
-  babel?: ConfigFunction;
-  autoprefixer?: ConfigFunction;
-  postcss?: ConfigFunction;
-  styledComponents?: ConfigFunction;
-  lodash?: ConfigFunction;
-  devServer?: DevServerConfig;
-  tsLoader?: ConfigFunction;
-  terser?: ConfigFunction;
-  minifyCss?: ConfigFunction;
-  esbuild?: Record<string, unknown>;
-}
-
-type RuntimeConfig = Record<string, any>;
-
-interface RuntimeByEntriesConfig {
-  [name: string]: RuntimeConfig;
-}
-
-interface UserConfig {
-  source?: SourceConfig;
-  output?: OutputConfig;
-  server?: ServerConfig;
-  dev?: DevConfig;
-  deploy?: DeployConfig;
-  tools?: ToolsConfig;
-  plugins?: PluginConfig;
-  runtime?: RuntimeConfig;
-  runtimeByEntries?: RuntimeByEntriesConfig;
-}
-
-type ConfigParam =
-  | UserConfig
-  | Promise<UserConfig>
-  | ((env: any) => UserConfig | Promise<UserConfig>);
-
-interface LoadedConfig {
-  config: UserConfig;
-  filePath: string | false;
-  dependencies: string[];
-  pkgConfig: UserConfig;
-  jsConfig: UserConfig;
-}
+export * from './types';
 
 export const defineConfig = (config: ConfigParam): ConfigParam => config;
 
@@ -253,7 +83,7 @@ const showAdditionalPropertiesError = (error: ErrorObject) => {
   }
 };
 
-/* eslint-disable max-statements, max-params */
+/* eslint-disable  max-statements, max-params */
 export const resolveConfig = async (
   loaded: LoadedConfig,
   configs: UserConfig[],
@@ -333,18 +163,5 @@ export const resolveConfig = async (
 
   return resolved;
 };
-/* eslint-enable max-statements, max-params */
 
-export type {
-  SourceConfig,
-  OutputConfig,
-  ServerConfig,
-  DevConfig,
-  DeployConfig,
-  ToolsConfig,
-  RuntimeConfig,
-  RuntimeByEntriesConfig,
-  UserConfig,
-  ConfigParam,
-  LoadedConfig,
-};
+/* eslint-enable max-statements, max-params */
