+++ packages/cli/plugin-proxy/package.json
@@ -17,16 +17,6 @@
   "main": "./dist/js/node/index.js",
   "module": "./dist/js/treeshaking/index.js",
   "jsnext:modern": "./dist/js/modern/index.js",
-  "typesVersions": {
-    "*": {
-      ".": [
-        "./dist/types/index.d.ts"
-      ],
-      "types": [
-        "./dist/types/types.d.ts"
-      ]
-    }
-  },
   "exports": {
     ".": {
       "node": {
@@ -39,14 +29,6 @@
     "./cli": {
       "jsnext:source": "./src/index.ts",
       "default": "./dist/js/node/index.js"
-    },
-    "./types": {
-      "node": {
-        "import": "./dist/js/modern/types.js",
-        "require": "./dist/js/node/types.js",
-        "types": "./dist/types/types.d.ts"
-      },
-      "default": "./dist/js/treeshaking/types.js"
     }
   },
   "scripts": {
