--- packages/cli/plugin-proxy/src/types.ts
@@ -1,9 +1,0 @@
-import '@modern-js/core';
-
-export type ProxyOptions = string | Record<string, string>;
-
-declare module '@modern-js/core' {
-  export interface DevConfig {
-    proxy?: ProxyOptions;
-  }
-}
