+++ packages/cli/plugin-tailwind/package.json
@@ -29,24 +29,6 @@
     "./cli": {
       "jsnext:source": "./src/cli.ts",
       "default": "./dist/js/node/cli.js"
-    },
-    "./types": {
-      "node": {
-        "import": "./dist/js/modern/types.js",
-        "require": "./dist/js/node/types.js",
-        "types": "./dist/types/types.d.ts"
-      },
-      "default": "./dist/js/treeshaking/types.js"
-    }
-  },
-  "typesVersions": {
-    "*": {
-      ".": [
-        "./dist/types/index.d.ts"
-      ],
-      "types": [
-        "./dist/types/types.d.ts"
-      ]
     }
   },
   "scripts": {
