+++ packages/cli/plugin-tailwind/src/index.ts
@@ -1,3 +1,1 @@
-import './types';
-
 export { default } from './cli';
