+++ packages/cli/plugin-unbundle/package.json
@@ -25,24 +25,6 @@
         "require": "./dist/js/node/index.js"
       },
       "default": "./dist/js/treeshaking/index.js"
-    },
-    "./types": {
-      "node": {
-        "import": "./dist/js/modern/types.js",
-        "require": "./dist/js/node/types.js",
-        "types": "./dist/types/types.d.ts"
-      },
-      "default": "./dist/js/treeshaking/types.js"
-    }
-  },
-  "typesVersions": {
-    "*": {
-      ".": [
-        "./dist/types/index.d.ts"
-      ],
-      "types": [
-        "./dist/types/types.d.ts"
-      ]
     }
   },
   "files": [
